Move Cosmos extensions into main AzureResourceExtensions file.
